+++ src/background/popupData/PopupData.ts
@@ -13,6 +13,7 @@ import { flagsStorage } from '../flagsStorage';
 import { exclusions } from '../exclusions';
 import { forwarder } from '../forwarder';
 import { rateModal } from '../rateModal';
+import { mobileEdgePromoService } from '../mobileEdgePromoService';
 import { type EndpointsInterface } from '../endpoints/Endpoints';
 import { type PermissionsCheckerInterface } from '../permissionsChecker/PermissionsChecker';
 import { type PermissionsErrorInterface } from '../permissionsChecker/permissionsError';
@@ -62,6 +63,7 @@ interface PopupDataInterface {
     }
     isVpnEnabledByUrl?: boolean;
     shouldShowRateModal?: boolean;
+    shouldShowMobileEdgePromoBanner?: boolean;
     username?: string | null;
     shouldShowHintPopup?: boolean;
     isHostPermissionsGranted: boolean;
@@ -153,6 +155,7 @@ export class PopupData {
         const shouldShowRateModal = rateModal.shouldShowRateModal();
         const username = await this.credentials.getUsername();
         const shouldShowHintPopup = await hintPopup.shouldShowHintPopup();
+        const shouldShowMobileEdgePromoBanner = await mobileEdgePromoService.shouldShowBanner();
 
         // If error check permissions when popup is opened, ignoring multiple retries
         if (error) {
@@ -187,6 +190,7 @@ export class PopupData {
             flagsStorageData,
             isVpnEnabledByUrl,
             shouldShowRateModal,
+            shouldShowMobileEdgePromoBanner,
             username,
             shouldShowHintPopup,
             isVpnBlocked,