+++ src/popup/components/GlobalError/GlobalError.tsx
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react';
 
 import { rootStore } from '../../stores';
 import { popupActions } from '../../actions/popupActions';
-import { reactTranslator } from '../../../common/reactTranslator';
+import { translator } from '../../../common/translator';
 import { isLocationsNumberAcceptable } from '../../../common/is-locations-number-acceptable';
 import { useTelemetryPageViewEvent } from '../../../common/telemetry';
 import { TelemetryScreenName } from '../../../background/telemetry';
@@ -25,11 +25,6 @@ export const GlobalError = observer(() => {
         CONTROL: 'control',
     };
 
-    const ICON_TYPES = {
-        ERROR: 'error',
-        TROUBLE: 'trouble',
-    };
-
     const handleTryAgain = async (): Promise<void> => {
         await settingsStore.checkPermissions();
         // forcibly update locations on try again
@@ -48,7 +43,6 @@ export const GlobalError = observer(() => {
     };
 
     let errorType = ERROR_TYPES.PERMISSION;
-    const descriptionClassName = 'global-error__description';
 
     if (settingsStore.hasGlobalError) {
         errorType = ERROR_TYPES.PERMISSION;
@@ -68,85 +62,62 @@ export const GlobalError = observer(() => {
         canSendTelemetry,
     );
 
-    const errorsMap = {
-        [ERROR_TYPES.CONTROL]: {
-            title: reactTranslator.getMessage('control_error_title'),
-            description: reactTranslator.getMessage('control_error_description'),
-            icon: ICON_TYPES.TROUBLE,
-            buttons: [
-                {
-                    id: 1,
-                    handler: handleDisableExtensions,
-                    className: 'button button--medium button--green global-error__button',
-                    text: reactTranslator.getMessage('control_error_disable'),
-                },
-            ],
-        },
-        [ERROR_TYPES.PERMISSION]: {
-            title: reactTranslator.getMessage('global_error_title'),
-            description: reactTranslator.getMessage('global_error_description'),
-            icon: ICON_TYPES.ERROR,
-            buttons: [
-                {
-                    id: 1,
-                    handler: handleLearnMore,
-                    text: reactTranslator.getMessage('global_error_learn_more'),
-                    className: 'button button--medium button--green global-error__button',
-                },
-                {
-                    id: 2,
-                    handler: handleTryAgain,
-                    className: 'button button--medium button--link global-error__button',
-                    text: reactTranslator.getMessage('global_error_try_again'),
-                },
-            ],
-        },
-    };
-
-    const {
-        title, description, buttons, icon,
-    } = errorsMap[errorType];
-
-    const renderButtons = () => {
-        return buttons.map((button) => {
-            const {
-                id,
-                handler,
-                className,
-                text,
-            } = button;
-
-            return (
-                <button
-                    key={id}
-                    type="button"
-                    className={className}
-                    onClick={handler}
-                >
-                    {text}
-                </button>
-            );
-        });
-    };
+    if (errorType === ERROR_TYPES.CONTROL) {
+        return (
+            <div className="new-global-error new-global-error--control">
+                <div className="new-global-error__image-wrapper">
+                    <img
+                        src="../../../assets/images/confused.svg"
+                        className="new-global-error__image"
+                        alt="Confused Ninja"
+                    />
+                </div>
+                <div className="new-global-error__content">
+                    <div className="new-global-error__title">
+                        {translator.getMessage('control_error_title')}
+                    </div>
+                    <div className="new-global-error__description">
+                        {translator.getMessage('control_error_description')}
+                    </div>
+                    <button
+                        type="button"
+                        onClick={handleDisableExtensions}
+                        className="button button--large button--green"
+                    >
+                        {translator.getMessage('control_error_disable')}
+                    </button>
+                </div>
+            </div>
+        );
+    }
 
     return (
         <div className="global-error">
-            {errorType === ERROR_TYPES.PERMISSION && (
-                <div className="global-error__pic" />
-            )}
+            <div className="global-error__pic" />
             <div className="global-error__content">
-                <div className={`global-error__icon global-error__icon--${icon}`} />
-                {title && (
-                    <div className="global-error__title">
-                        {title}
-                    </div>
-                )}
-                <div className={descriptionClassName}>
-                    {description}
+                <div className="global-error__icon" />
+                <div className="global-error__title">
+                    {translator.getMessage('global_error_title')}
+                </div>
+                <div className="global-error__description">
+                    {translator.getMessage('global_error_description')}
                 </div>
             </div>
             <div className="global-error__actions">
-                {renderButtons()}
+                <button
+                    type="button"
+                    className="button button--medium button--green global-error__button"
+                    onClick={handleLearnMore}
+                >
+                    {translator.getMessage('global_error_learn_more')}
+                </button>
+                <button
+                    type="button"
+                    className="button button--medium button--link global-error__button"
+                    onClick={handleTryAgain}
+                >
+                    {translator.getMessage('global_error_try_again')}
+                </button>
             </div>
         </div>
     );