+++ src/popup/stores/GlobalStore.ts
@@ -75,6 +75,7 @@ export class GlobalStore {
                 isVpnEnabledByUrl,
                 shouldShowRateModal,
                 shouldShowHintPopup,
+                shouldShowMobileEdgePromoBanner,
                 isVpnBlocked,
                 isHostPermissionsGranted,
                 locationsTab,
@@ -82,8 +83,6 @@ export class GlobalStore {
             } = popupData;
 
             settingsStore.setForwarderDomain(forwarderDomain);
-
-            authStore.setIsAuthenticated(isAuthenticated);
 
             telemetryStore.setIsHelpUsImproveEnabled(helpUsImprove);
 
@@ -127,13 +126,13 @@ export class GlobalStore {
             settingsStore.setConnectivityState(connectivityState);
             settingsStore.setIsRoutable(isRoutable);
             settingsStore.setIsVpnBlocked(isVpnBlocked);
+            settingsStore.setShowMobileEdgePromoBanner(shouldShowMobileEdgePromoBanner);
             settingsStore.setLimitedOfferData(limitedOfferData);
             settingsStore.setPromoNotification(promoNotification);
             vpnStore.setVpnInfo(vpnInfo);
             vpnStore.setLocations(locations);
             vpnStore.setSelectedLocation(selectedLocation);
             vpnStore.setIsPremiumToken(isPremiumToken);
-            authStore.setIsAuthenticated(isAuthenticated);
             await authStore.getAuthCacheFromBackground();
             await authStore.setPolicyAgreement(policyAgreement);
             await settingsStore.checkRateStatus();
@@ -148,6 +147,19 @@ export class GlobalStore {
         }
     }
 
+    /**
+     * Retrieves the authentication status from the
+     * background and marks the status as retrieved.
+     */
+    @action
+    async initAuthenticatedStatus(): Promise<void> {
+        const { authStore } = this.rootStore;
+
+        const result = await messenger.isAuthenticated();
+        authStore.setIsAuthenticated(!!result);
+        authStore.setAuthenticatedStatusRetrieved(true);
+    }
+
     @action
     async init(): Promise<void> {
         /**
@@ -155,6 +167,13 @@ export class GlobalStore {
          * and UI might shift because it was loaded too late.
          */
         await this.getAndroidData();
+
+        /**
+         * Authentication status should be retrieved from background before
+         * popup data is retrieved to prevent flickering of the loaders.
+         */
+        await this.initAuthenticatedStatus();
+
         await this.getPopupData(MAX_GET_POPUP_DATA_ATTEMPTS);
         await this.getDesktopAppData();
         await this.rootStore.authStore.getResendCodeCountdownAndStart();