+++ tasks/edge/webpack.edge.ts
@@ -18,8 +18,6 @@ import {
 
 import { edgeManifestDiff } from './manifest.edge';
 
-const BACKGROUND_PATH = path.resolve(__dirname, '..', SRC_PATH, 'background');
-
 const EDGE_PATH = 'edge';
 
 let zipFilename = 'edge.zip';
@@ -30,19 +28,30 @@ if (IS_DEV && STAGE_ENV === StageEnv.Prod) {
 
 const commonConfig = getCommonConfig(Browser.Edge);
 
+const OFFSCREEN_PATH = path.resolve(__dirname, '..', SRC_PATH, 'offscreen');
+const WORKER_SCRIPT = path.resolve(__dirname, '..', SRC_PATH, 'worker/worker.ts');
+
 const plugins: webpack.WebpackPluginInstance[] = [
-    // TODO: on move to MV3 inject Mv3Timers
+    new HtmlWebpackPlugin({
+        template: path.join(OFFSCREEN_PATH, 'index.html'),
+        filename: 'offscreen.html',
+        chunks: ['offscreen'],
+        cache: false,
+    }),
     new webpack.NormalModuleReplacementPlugin(/\.\/AbstractTimers/, ((resource: any) => {
+        // TODO remove this replacement when MV3 will fix alarms bug,
+        //  https://github.com/AdguardTeam/AdGuardVPNExtension/issues/116
+        //  https://bugs.chromium.org/p/chromium/issues/detail?id=1472759
         // eslint-disable-next-line no-param-reassign
         resource.request = resource.request.replace(/\.\/AbstractTimers/, './Mv2Timers');
     })),
     new webpack.NormalModuleReplacementPlugin(/\.\/networkConnectionObserverAbstract/, ((resource: any) => {
         // eslint-disable-next-line no-param-reassign
-        resource.request = resource.request.replace(/\.\/networkConnectionObserverAbstract/, './networkConnectionObserverMv2');
+        resource.request = resource.request.replace(/\.\/networkConnectionObserverAbstract/, './networkConnectionObserverMv3');
     })),
     new webpack.NormalModuleReplacementPlugin(/\.\/abstractProxyAuthTrigger/, ((resource: any) => {
         // eslint-disable-next-line no-param-reassign
-        resource.request = resource.request.replace(/\.\/abstractProxyAuthTrigger/, './mv2');
+        resource.request = resource.request.replace(/\.\/abstractProxyAuthTrigger/, './mv3');
     })),
     new CopyWebpackPlugin({
         patterns: [
@@ -52,12 +61,6 @@ const plugins: webpack.WebpackPluginInstance[] = [
                 transform: (content: Buffer) => updateManifest(content, edgeManifestDiff),
             },
         ],
-    }),
-    new HtmlWebpackPlugin({
-        template: path.join(BACKGROUND_PATH, 'index.html'),
-        filename: 'background.html',
-        chunks: ['background'],
-        cache: false,
     }),
     new ZipWebpackPlugin({
         path: '../',
@@ -72,6 +75,10 @@ if (!outputPath) {
 }
 
 const edgeDiffConfig = {
+    entry: {
+        offscreen: OFFSCREEN_PATH,
+        worker: WORKER_SCRIPT,
+    },
     output: {
         path: path.join(outputPath, EDGE_PATH),
     },